Split up ASMPatches, move asm from LuaInterfaces
