+++ repentogon/LuaInterfaces/Entities/EntityNPC.h
@@ -1,4 +1,1 @@
 #pragma once
-
-void PatchFireProjectiles();
-void PatchFireBossProjectiles();
