+++ repentogon/LuaInterfaces/Entities/LuaEntityPlayer.cpp
@@ -1,12 +1,9 @@
-﻿#include <algorithm>
-#include <random>
-
-#include "IsaacRepentance.h"
+﻿#include "IsaacRepentance.h"
 #include "LuaCore.h"
-#include "ASMPatcher.hpp"
-#include "SigScan.h"
 #include "HookSystem.h"
+
 #include "../LuaWeapon.h"
+#include "../../Patches/ASMPatches/ASMPlayer.h"
 
 /*
 
@@ -22,27 +19,7 @@
  (copycat of kilburn's ascii art of Isaac)
 */
 
-
-// this is from LuaInit but with a range of 0, 1 instead of -1, 1
-static std::uniform_real_distribution<float> _distrib(0, 1);
-static std::random_device rd;
-static std::mt19937 gen(rd());
-
 std::map<int, int> fakeItems;
-
-struct CheckFamiliarStorage {
-	std::vector<Entity_Familiar*> familiars;
-	bool inUse = false;
-};
-
-thread_local CheckFamiliarStorage familiarsStorage;
-
-static std::vector<Entity_Familiar*>& InitFamiliarStorage() {
-	std::vector<Entity_Familiar*>& familiars = familiarsStorage.familiars;
-	familiars.clear();
-	familiarsStorage.inUse = true;
-	return familiars;
-}
 
 LUA_FUNCTION(Lua_GetMultiShotPositionVelocity) // This *should* be in the API, but magically vanished some point after 1.7.8.
 {
@@ -1060,7 +1037,8 @@ LUA_FUNCTION(Lua_SpawnAquariusCreep) {
 	}
 	else
 	{
-		effect->_sprite._scale *= ((_distrib(gen) * 0.5f) + 0.2f);
+		float random = (static_cast <float> (rand()) / static_cast <float> (RAND_MAX));
+		effect->_sprite._scale *= ((random * 0.5f) + 0.2f);
 		effect->_collisionDamage = params._tearDamage;
 		effect->SetColor(&params._tearColor, 0, -1, true, false);
 
@@ -1322,32 +1300,6 @@ LUA_FUNCTION(Lua_EntityPlayer_CheckFamiliarEx) {
 	FamiliarStorageToLua(L, familiars);
 
 	return 1;
-}
-
-void __stdcall CheckFamiliar_Internal(Entity_Familiar* familiar) {
-	familiar->Update();
-
-	if (familiarsStorage.inUse) {
-		familiarsStorage.familiars.push_back(familiar);
-	}
-	
-	return;
-}
-
-void PatchCheckFamiliar() {
-	SigScan scanner("8b06ff50??8b75"); // this is immediately before the call to Update()
-	scanner.Scan();
-	void* addr = scanner.GetAddress();
-
-	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS, true);
-	ASMPatch patch;
-	patch.PreserveRegisters(savedRegisters)
-		.Push(ASMPatch::Registers::ESI)  // Push the familiar spawned
-		.AddInternalCall(CheckFamiliar_Internal) 
-		.RestoreRegisters(savedRegisters)
-		// this should neatly fit in the 5 bytes used to call Update, and we handle calling it there, nothing to restore
-		.AddRelativeJump((char*)addr + 0x5);
-	sASMPatcher.PatchAt(addr, &patch);
 }
 
 LUA_FUNCTION(Lua_PlayerGetEveSumptoriumCharge) {
