+++ repentogon/Patches/ASMPatches.h
@@ -3,11 +3,14 @@
 #include <bitset>
 
 #include "IsaacRepentance.h"
+#include "SigScan.h"
+#include "LuaCore.h"
+#include "Log.h"
+#include "../REPENTOGONOptions.h"
 
 extern std::bitset<500> CallbackState;
 extern int ambushWaves;
 extern bool overrideMegaSatanEnding;
 
 void __stdcall LogMessageCallback(const char* logMessage);
-bool __stdcall RunPreLaserCollisionCallback(Entity_Laser* laser, Entity* entity);
 void PerformASMPatches();
