+++ repentogon/Patches/ASMPatches/ASMCallbacks.cpp
@@ -1,0 +1,563 @@
+#include "IsaacRepentance.h"
+#include "SigScan.h"
+#include "LuaCore.h"
+
+#include "ASMPatcher.hpp"
+
+#include "../ASMPatches.h"
+#include "../XMLData.h"
+
+/* * MC_PRE_LASER_COLLISION * *
+* There is no function you can hook that would allow you to skip a laser "collision".
+* This section patches in PRE_LASER_COLLISION callbacks right when the game would
+* first consider an entity to be hit by the laser. Two patches are needed because
+* "sample" lasers (lasers that are capable of curving) use different collision code.
+* POST_LASER_COLLISION is just done as a function hook in CustomCallbacks.cpp
+*/
+
+bool __stdcall RunPreLaserCollisionCallback(Entity_Laser* laser, Entity* entity) {
+	const int callbackid = 1248;
+	if (CallbackState.test(callbackid - 1000)) {
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+		lua::LuaResults result = lua::LuaCaller(L).push(callbackid)
+			.push(*laser->GetVariant())
+			.push(laser, lua::Metatables::ENTITY_LASER)
+			.push(entity, lua::Metatables::ENTITY)
+			.call(1);
+
+		if (!result) {
+			if (lua_isboolean(L, -1)) {
+				return lua_toboolean(L, -1);
+			}
+		}
+	}
+	return false;
+}
+
+// This patch injects the PRE_LASER_COLLISION callback for "sample" lasers (lasers that curve, ie have multiple sample points).
+void PatchPreSampleLaserCollision() {
+	SigScan scanner("8b4d??8b41??83f809");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+
+	ASMPatch::SavedRegisters reg(ASMPatch::SavedRegisters::GP_REGISTERS_STACKLESS, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(reg)
+		.AddBytes("\x8B\x4D\x90")  // mov ecx,dword ptr ss:[ebp-70] (Put the entity in ECX)
+		.AddBytes("\x8B\x45\x8C")  // mov eax,dword ptr ss:[ebp-74] (Put the laser in EAX)
+		.AddBytes("\x51\x50") // Push ECX, EAX for function inputs
+		.AddInternalCall(RunPreLaserCollisionCallback) // Run PRE_LASER_COLLISION callback
+		.AddBytes("\x84\xC0") // TEST AL, AL
+		.RestoreRegisters(reg)
+		.AddConditionalRelativeJump(ASMPatcher::CondJumps::JNZ, (char*)addr - 0x4B)  // Ignore collision
+		.AddBytes("\x8B\x4D\x90")  // mov ecx,dword ptr ss:[ebp-70] (Put the entity in ECX) (Restored overridden command)
+		.AddBytes("\x8B\x41\x28")  // mov eax,dword ptr ds:[ecx+28] (Put the entity's type in EAX) (Restored overridden command)
+		.AddRelativeJump((char*)addr + 0x6);  // Allow collision
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// This patch injects the PRE_LASER_COLLISION callback for "non-sample" lasers (ones that can ONLY ever be a straight line).
+void PatchPreLaserCollision() {
+	SigScan scanner("f30f1095????????8bcf56");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+
+	ASMPatch::SavedRegisters reg(ASMPatch::SavedRegisters::GP_REGISTERS_STACKLESS, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(reg)
+		.AddBytes("\x56\x57") // Push EDI (entity) and ESI (laser) for function input
+		.AddInternalCall(RunPreLaserCollisionCallback) // Run PRE_LASER_COLLISION callback
+		.AddBytes("\x84\xC0") // TEST AL, AL
+		.RestoreRegisters(reg)
+		.AddConditionalRelativeJump(ASMPatcher::CondJumps::JNZ, (char*)addr + 0x49A) // Ignore collision
+		.AddBytes("\xF3\x0F\x10\x95\x34\xFF\xFF\xFF")  // movss xmm2,dword ptr ss:[EBP-0xCC] (Restored overridden command)
+		.AddRelativeJump((char*)addr + 0x8); // Allow collision
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// End of PRE_LASER_COLLISION patches
+
+/* * MC_ENTITY_TAKE_DMG REIMPLEMENTATION * *
+* This section patches in a new ENTITY_TAKE_DMG callback with table returns for overridding.
+* A patch is used because ENTITY_TAKE_DMG runs in the middle of the TakeDamage functions,
+* not right at the start, so we can't just run the callback to modify the function inputs.
+* By patching over where the callback would normally be triggered, we have access to the
+* TakeDamage function parameters in memory and can modify them directly.
+* We need to patch into both Entity::TakeDamage AND EntityPlayer::TakeDamage.
+*/
+extern int entityTakeDmgCallbackKey;
+
+bool __stdcall ProcessPreDamageCallback(Entity* entity, char* ebp, bool isPlayer) {
+	int callbackid = 1007;
+	if (CallbackState.test(callbackid - 1000)) {
+		// Obtain inputs as offsets from EBP (same way the compiled code reads them).
+		// As pointers so we can modify them :)
+		unsigned __int64* damageFlags = (unsigned __int64*)(ebp + 0x0C);
+		float* damage = (float*)(ebp + 0x08);
+		int* damageHearts = isPlayer ? (int*)(ebp - 0x100) : nullptr;
+		EntityRef** source = (EntityRef**)(ebp + 0x14);
+		int* damageCountdown = (int*)(ebp + 0x18);
+
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, entityTakeDmgCallbackKey);
+
+		unsigned int entityType = *entity->GetType();
+
+		lua::LuaCaller caller(L);
+		caller.push(callbackid)
+			.push(entityType)
+			.push(entity, lua::Metatables::ENTITY);
+		if (isPlayer) {
+			caller.push(*damageHearts);
+		}
+		else {
+			caller.push(*damage);
+		}
+		caller.push(*damageFlags)
+			.push(*source, lua::Metatables::ENTITY_REF)
+			.push(*damageCountdown);
+		lua::LuaResults lua_result = caller.call(1);
+
+		if (!lua_result) {
+			if (lua_isboolean(L, -1)) {
+				return lua_toboolean(L, -1);
+			}
+			else if (lua_istable(L, -1)) {
+				unsigned __int64 originalDamageFlags = *damageFlags;
+				// New table return format for overriding values.
+				lua_pushnil(L);
+				while (lua_next(L, -2) != 0) {
+					if (lua_isstring(L, -2)) {
+						const std::string key = lua_tostring(L, -2);
+						if (key == "Damage" && lua_isnumber(L, -1)) {
+							float newDamage = (float)lua_tonumber(L, -1);
+							if (newDamage < 0) {
+								newDamage = 0;
+							}
+							if (isPlayer) {
+								*damageHearts = (int)newDamage;
+							}
+							*damage = newDamage;
+						}
+						else if (key == "DamageFlags" && lua_isinteger(L, -1)) {
+							*damageFlags = lua_tointeger(L, -1);
+						}
+						// Couldn't get this to work. Maybe garbage collection related?
+						/*else if (key == "Source" && lua_isuserdata(L, -1)) {
+							*source = &lua::GetUserdata<EntityRef>(L, -1, lua::Metatables::ENTITY_REF, "EntityRef");
+							void* newSource = lua::CheckUserdata(L, -1, lua::Metatables::ENTITY_REF, "EntityRef");
+							if (newSource != nullptr) {
+								*source = (EntityRef*)newSource;
+							}
+						}*/
+						else if (key == "DamageCountdown" && lua_isnumber(L, -1)) {
+							*damageCountdown = (int)lua_tonumber(L, -1);
+						}
+					}
+					lua_pop(L, 1);
+				}
+
+				if (!isPlayer) {
+					// If certain DamageFlags were added/present, retroactively obey their behaviour.
+					// (Normally these are handled prior to when this callback runs.)
+
+					// Check if DamageFlag.DAMAGE_FIRE was added
+					if ((originalDamageFlags & (1 << 1)) < (*damageFlags & (1 << 1))) {
+						if (*entity->GetFireDamageCountdown() > 0) {
+							return false;
+						}
+						*entity->GetFireDamageCountdown() = 10;
+					}
+					// Check if DamageFlag.DAMAGE_COUNTDOWN was added
+					if ((originalDamageFlags & (1 << 6)) < (*damageFlags & (1 << 6))) {
+						if (*entity->GetDamageCountdown() > 0) {
+							return false;
+						}
+					}
+					// If the DAMAGE_COUNTDOWN flag is currently present, set the entity's
+					// DamageCountdown to the now potentially-modified value.
+					if ((*damageFlags & (1 << 6)) != 0) {
+						*entity->GetDamageCountdown() = *damageCountdown;
+					}
+				}
+			}
+		}
+	}
+
+	return true;
+}
+
+void InjectPreDamageCallback(void* addr, bool isPlayer) {
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS - ASMPatch::SavedRegisters::Registers::EAX, true);
+	ASMPatch patch;
+	patch.AddBytes("\x58\x58\x58\x58\x58")  // Pop EAX x5 to remove the overridden function's inputs from the stack.
+		.PreserveRegisters(savedRegisters)
+		.AddBytes("\x31\xC0");  // xor eax,eax
+	if (isPlayer) {
+		patch.AddBytes("\xB0\x01");  // Move Al, 0x1
+	}
+	patch.Push(ASMPatch::Registers::EAX)  // Push a boolean to identify if this is EntityPlayer:TakeDamge
+		.Push(ASMPatch::Registers::EBP)  // Push EBP (We'll get other inputs as offsets from this)
+		.Push(ASMPatch::Registers::EDI)  // Push Entity pointer
+		.AddInternalCall(ProcessPreDamageCallback);  // Run the new MC_ENTITY_TAKE_DMG callback
+	if (isPlayer) {
+		// For the player, also copy the (potentially modified) lower DamageFlags to [EBP-0x104]
+		patch.AddBytes("\x8B\x75\x0C")  // mov esi,dword ptr [EBP+0x0C]
+			.AddBytes("\x89\xB5\xFC\xFE\xFF\xFF");  // mov dword ptr [EBP-0x104],esi
+	}
+	patch.RestoreRegisters(savedRegisters);
+	if (!isPlayer) {
+		// For non-players, also copy the (potentially modified) DamageFlags into EBX and ESI
+		patch.AddBytes("\x8B\x5D\x0C")  // mov ebx,dword ptr [EBP+0x0C]
+			.AddBytes("\x8B\x75\x10");  // mov esi,dword ptr [EBP+0x10]
+	}
+	patch.AddRelativeJump((char*)addr + 0x5);
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// Patches overtop the existing calls to LuaEngine::Callback::EntityTakeDamage,
+// within the Entity::TakeDamage and EntityPlayer::TakeDamage functions respectively.
+void PatchPreEntityTakeDamageCallbacks() {
+	SigScan entityTakeDmgScanner("e8????????84c00f84????????8d4c24??e8????????f30f1045");
+	entityTakeDmgScanner.Scan();
+	InjectPreDamageCallback(entityTakeDmgScanner.GetAddress(), false);
+
+	SigScan playerTakeDmgScanner("e8????????84c00f84????????8bcfe8????????84c00f85????????83bf????????25");
+	playerTakeDmgScanner.Scan();
+	InjectPreDamageCallback(playerTakeDmgScanner.GetAddress(), true);
+}
+
+// End of ENTITY_TAKE_DMG patches
+
+/* * MC_POST_ENTITY_TAKE_DMG * *
+* A hook worked just fine for ""POST_TAKE_DMG"" initially, but now that we support overidding
+* values via ENTITY_TAKE_DMG (see above) we need the POST_TAKE_DMG callback to actually show
+* those modified values. That means another patch!!
+* Like before, we need to patch into both Entity::TakeDamage and EntityPlayer::TakeDamage.
+*/
+
+void __stdcall ProcessPostDamageCallback(Entity* entity, char* ebp, bool isPlayer) {
+	int callbackid = 1006;
+	if (CallbackState.test(callbackid - 1000)) {
+		// Obtain inputs as offsets from EBP (same way the compiled code reads them).
+		unsigned __int64 damageFlags = *(unsigned __int64*)(ebp + 0x0C);
+		float damage = *(float*)(ebp + 0x08);
+		int damageHearts = isPlayer ? *(int*)(ebp - 0x100) : 0;
+		EntityRef* source = *(EntityRef**)(ebp + 0x14);
+		int damageCountdown = *(int*)(ebp + 0x18);
+
+		if (isPlayer && source->_type == 33 && source->_variant == 4) {
+			// The white fireplace is a unique case where the game considers the player to have taken "damage"
+			// but no on-damage effets are triggered. Don't run the post-damage callback in this case.
+			return;
+		}
+
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+		lua::LuaCaller caller(L);
+		caller.push(callbackid)
+			.push(*entity->GetType())
+			.push(entity, lua::Metatables::ENTITY);
+		if (isPlayer) {
+			caller.push(damageHearts);
+		}
+		else {
+			caller.push(damage);
+		}
+		caller.push(damageFlags)
+			.push(source, lua::Metatables::ENTITY_REF)
+			.push(damageCountdown)
+			.call(1);
+	}
+};
+
+void InjectPostDamageCallback(void* addr, bool isPlayer) {
+	const int numOverriddenBytes = (isPlayer ? 10 : 5);
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS, true);
+	ASMPatch patch;
+	patch.AddBytes(ByteBuffer().AddAny((char*)addr, numOverriddenBytes));  // Restore the commands we overwrote
+	if (isPlayer) {
+		// The EntityPlayer::TakeDamage patch needs to ask Al if the damage occured.
+		// The Entity::TakeDamage patch doesn't need to do this since it is at a location that only runs after damage.
+		patch.AddBytes("\x84\xC0") // Test Al (Al?)
+			.AddConditionalRelativeJump(ASMPatcher::CondJumps::JZ, (char*)addr + numOverriddenBytes);
+	}
+	patch.PreserveRegisters(savedRegisters)
+		.AddBytes("\x31\xC0");  // xor eax,eax
+	if (isPlayer) {
+		patch.AddBytes("\xB0\x01");  // Move Al, 0x1
+	}
+	patch.Push(ASMPatch::Registers::EAX)  // Push a boolean to identify if this is EntityPlayer:TakeDamge
+		.Push(ASMPatch::Registers::EBP)  // Push EBP (We'll get other inputs as offsets from this)
+		.Push(ASMPatch::Registers::EDI)  // Push Entity pointer
+		.AddInternalCall(ProcessPostDamageCallback)  // Run the MC_POST_(ENTITY_)TAKE_DMG callback
+		.RestoreRegisters(savedRegisters)
+		.AddRelativeJump((char*)addr + numOverriddenBytes);
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// These patches overwrite suitably-sized commands near where the respective TakeDamage functions would return.
+// The overridden bytes are restored by the patch.
+void PatchPostEntityTakeDamageCallbacks() {
+	SigScan entityTakeDmgScanner("f30f1147??5f5e5b8be55dc21400");
+	entityTakeDmgScanner.Scan();
+	InjectPostDamageCallback(entityTakeDmgScanner.GetAddress(), false);
+
+	SigScan playerTakeDmgScanner("8b4d??64890d????????595f5e8b4d??33cde8????????8be55dc2140068????????e8????????83c404833d????????ff0f85????????c745??0d000000");
+	playerTakeDmgScanner.Scan();
+	InjectPostDamageCallback(playerTakeDmgScanner.GetAddress(), true);
+}
+
+// End of POST_ENTITY_TAKE_DMG patches
+
+
+// MC_PRE_PLAYER_USE_BOMB
+bool __stdcall ProcessPrePlayerUseBombCallback(Entity_Player* player) {
+	int callbackid = 1220;
+	if (CallbackState.test(callbackid - 1000)) {
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+		lua::LuaResults lua_result = lua::LuaCaller(L).push(callbackid)
+			.push(*player->GetVariant())
+			.push(player, lua::Metatables::ENTITY_PLAYER)
+			.call(1);
+
+		if (!lua_result && lua_isboolean(L, -1)) {
+			return lua_toboolean(L, -1);
+		}
+	}
+	return true;
+}
+
+void ASMPatchPrePlayerUseBomb() {
+	SigScan scanner("6a0068050200008bcf");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(savedRegisters)
+		.Push(ASMPatch::Registers::EDI)  // Push the player
+		.AddInternalCall(ProcessPrePlayerUseBombCallback)  // Run MC_PRE_PLAYER_USE_BOMB
+		.AddBytes("\x84\xC0") // TEST AL, AL
+		.RestoreRegisters(savedRegisters)
+		.AddConditionalRelativeJump(ASMPatcher::CondJumps::JE, (char*)addr + 0x7C5) // Jump for false (negate the bomb placement)
+		.AddBytes(ByteBuffer().AddAny((char*)addr, 0x7))  // Restore the 7 bytes we overwrote
+		.AddRelativeJump((char*)addr + 0x7);  // Jump for true (allow the bomb placement)
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// MC_POST_PLAYER_USE_BOMB
+void __stdcall ProcessPostPlayerUseBombCallback(Entity_Player* player, Entity_Bomb* bomb) {
+	int callbackid = 1221;
+	if (CallbackState.test(callbackid - 1000)) {
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+		lua::LuaCaller(L).push(callbackid)
+			.push(*player->GetVariant())
+			.push(player, lua::Metatables::ENTITY_PLAYER)
+			.push(bomb, lua::Metatables::ENTITY_BOMB)
+			.call(0);
+	}
+}
+
+void ASMPatchPostPlayerUseBomb() {
+	SigScan scanner("8b7424??46897424??3b7424??0f8c");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(savedRegisters)
+		.Push(ASMPatch::Registers::ESI)  // Push the BOMB
+		.Push(ASMPatch::Registers::EDI)  // Push the PLAYER
+		.AddInternalCall(ProcessPostPlayerUseBombCallback)  // Run MC_POST_PLAYER_USE_BOMB
+		.RestoreRegisters(savedRegisters)
+		.AddBytes(ByteBuffer().AddAny((char*)addr, 0x5))  // Restore the 5 bytes we overwrote
+		.AddRelativeJump((char*)addr + 0x5);
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+// MC_PRE_M_MORPH_ACTIVE
+int __stdcall RunPreMMorphActiveCallback(Entity_Player* player, int collectibleId) {
+	int callbackId = 1190;
+	if (CallbackState.test(callbackId - 1000)) {
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+		lua::LuaResults result = lua::LuaCaller(L).push(callbackId)
+			.pushnil()
+			.push(player, lua::Metatables::ENTITY_PLAYER)
+			.push(collectibleId)
+			.call(1);
+
+		if (!result) {
+			if (lua_isboolean(L, -1)) {
+				if (!lua_toboolean(L, -1)) {
+					return 0;
+				}
+			}
+			else if (lua_isinteger(L, -1)) {
+				return (int)lua_tointeger(L, -1);
+			}
+		}
+	}
+
+	return collectibleId;
+}
+
+// MC_PRE_M_MORPH_ACTIVE
+// This callback triggers when an active gets rerolled by 'M (trinket id 138) and allows for overriding its behavior.
+void ASMPatchPreMMorphActiveCallback() {
+	SigScan scanner("85f674??6a016a00");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+
+	printf("[REPENTOGON] Patching Entity_Player::TriggerActiveItemUsed at %p\n", addr);
+
+	ASMPatch::SavedRegisters registers(ASMPatch::SavedRegisters::Registers::GP_REGISTERS_STACKLESS - ASMPatch::SavedRegisters::Registers::ESI, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(registers)
+		.Push(ASMPatch::Registers::ESI) // push the item id
+		.Push(ASMPatch::Registers::EDI) // push the player
+		.AddInternalCall(RunPreMMorphActiveCallback) // run MC_PRE_M_MORPH_ACTIVE
+		.AddBytes("\x89\xC6") // mov esi, eax
+		.AddBytes("\x85\xF6") // test esi, esi
+		.RestoreRegisters(registers)
+		.AddConditionalRelativeJump(ASMPatcher::CondJumps::JZ, (char*)addr + 0x27) // jump for 0 (don't reroll the active)
+		.AddBytes(ByteBuffer().AddAny((char*)addr + 4, 0x2)) // restore the instruction we overwrote
+		.AddRelativeJump((char*)addr + 0x6); // jump for everything else (reroll the active)
+
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+bool resSplit = false;
+void __stdcall TrySplitTrampoline(Entity_NPC* npc, bool result) {
+	resSplit = result;
+
+	if (npc != nullptr) {
+		int callbackid = 1191;
+
+		if (CallbackState.test(callbackid - 1000)) {
+			lua_State* L = g_LuaEngine->_state;
+			lua::LuaStackProtector protector(L);
+
+			lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+			lua::LuaResults result = lua::LuaCaller(L).push(callbackid)
+				.push(npc->_type)
+				.push(npc, lua::Metatables::ENTITY_NPC)
+				.push(resSplit)
+
+				.call(1);
+
+			if (!result) {
+				if (lua_isboolean(L, -1)) {
+					resSplit = lua_toboolean(L, -1);
+					return;
+				}
+			}
+		}
+		XMLAttributes xmlData = XMLStuff.EntityData->GetNodesByTypeVarSub(npc->_type, npc->_variant, npc->_subtype, false);
+		const std::string nosplit = xmlData["nosplit"];
+
+		if (nosplit == "true") {
+			resSplit = true;
+		}
+		else if (nosplit == "false") {
+			resSplit = false;
+		}
+	}
+}
+
+void ASMPatchTrySplit() {
+	SigScan scanner("f30f100d????????0f2f8b");
+	scanner.Scan();
+	void* addr = scanner.GetAddress();
+	printf("[REPENTOGON] Patching Entity_NPC::TrySplit at %p\n", addr);
+	void* ptr = &resSplit;
+	const int numOverriddenBytes = 8;
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS + ASMPatch::SavedRegisters::Registers::XMM0 + ASMPatch::SavedRegisters::Registers::XMM1, true);
+	ASMPatch patch;
+	patch.PreserveRegisters(savedRegisters);
+	patch.AddBytes("\x0f\xb6\xc0") // MOVZX EAX,AL
+		.Push(ASMPatch::Registers::EAX)  // Push current result
+		.Push(ASMPatch::Registers::EBX)  // Push NPC
+		.AddInternalCall(TrySplitTrampoline)
+		.RestoreRegisters(savedRegisters)
+		.AddBytes("\xA0").AddBytes(ByteBuffer().AddAny((char*)&ptr, 4)) // mov al, byte ptr ds:[XXXXXXXX]
+		.AddBytes(ByteBuffer().AddAny((char*)addr, numOverriddenBytes))  // Restore the commands we overwrote
+		.AddRelativeJump((char*)addr + numOverriddenBytes);
+	sASMPatcher.PatchAt(addr, &patch);
+}
+
+void ASMPatchInputAction()
+{
+	const char* signatures[] = {
+		"837f??0275??833d????????0074??8d45??505351",
+		"8378??0275??833d????????00",
+		"837f??0275??833d????????0074??8d45??50536a00",
+		"837f??0275??833d????????0074??8d45??50536a01",
+		NULL
+	};
+
+	const char** start = signatures;
+	while (*start) {
+		const char* signature = *start;
+		SigScan scanner(signature);
+		if (!scanner.Scan()) {
+			ZHL::Log("[fatal] ASMPatchInputAction: could not find signature %s\n", signature);
+			throw std::runtime_error("Could not find signature");
+		}
+		else {
+			ZHL::Log("[info] ASMPatchInputAction: found signature at %p\n", scanner.GetAddress());
+		}
+
+		ASMPatch patch(ByteBuffer().AddByte(0x90, 6));
+		sASMPatcher.FlatPatch(scanner.GetAddress(), &patch);
+
+		++start;
+	}
+}
+
+void __stdcall RunPostNightmareSceneRender(NightmareScene* ns) {
+	const int callbackid = 1102;
+
+	if (CallbackState.test(callbackid - 1000)) {
+		lua_State* L = g_LuaEngine->_state;
+		lua::LuaStackProtector protector(L);
+		lua_rawgeti(L, LUA_REGISTRYINDEX, g_LuaEngine->runCallbackRegistry->key);
+
+		lua::LuaResults result = lua::LuaCaller(L).push(callbackid)
+			.pushnil()
+			//.push(ns, lua::metatables::NightmareSceneMT)
+			.call(1);
+	}
+}
+
+void ASMPatchPostNightmareSceneCallback() {
+	ASMPatch::SavedRegisters savedRegisters(ASMPatch::SavedRegisters::Registers::GP_REGISTERS, true);
+	ASMPatch patch;
+
+	SigScan scanner_transition("f30f108f????????0f57c00f2fc80f86????????f30f1005");
+	scanner_transition.Scan();
+	void* addr = scanner_transition.GetAddress();
+
+	patch.PreserveRegisters(savedRegisters)
+		.Push(ASMPatch::Registers::EDI) // NightmareScene
+		.AddInternalCall(RunPostNightmareSceneRender)
+		.RestoreRegisters(savedRegisters)
+		.AddBytes(ByteBuffer().AddAny((char*)addr, 8))  // Restore the commands we overwrote
+		.AddRelativeJump((char*)addr + 8);
+	sASMPatcher.PatchAt(addr, &patch);
+}
