+++ repentogon/Patches/ASMPatches/ASMDelirium.h
