+++ repentogon/Patches/ASMPatches/ASMMenu.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void ASMPatchModsMenu();
+void ASMPatchMenuOptionsLanguageChange();
