+++ repentogon/Patches/ASMPatches/ASMRoom.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void ASMPatchAmbushWaveCount();
+void ASMPatchMegaSatanEnding();
+void PatchRoomClearDelay();
